This is built from mysystem_sc commit: df654bb
This release changes how nodeStates are created.
Now new nodeStates are only created when the diagram is 'submitted' or 'checked', and when
the step is left.

It also fixes some bugs around the data not being posted right away when it should.

And finally it adds a submit count to the learner data.