Built from MySystem (e894b83): fixes author saving bug.
The previous commit to this repo was bad, because it was based on some
experimental refactoring of the authoring tool.

I need to update the javascript rule but in the advanced mode setting, it is too complicated for me to change. I don't want to mess up since students will be working on the project today.

Here's the javascript rule I want to add. Could you replace it with the old one? Project ID 4659, Step 3.7